+++ src/main/java/lotto/domain/Lotto.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
-import static lotto.utils.NumberValidation.validateDuplicateNumber;
-import static lotto.utils.NumberValidation.validateNumberRange;
-import static lotto.utils.NumberValidation.validateNumberSize;
+import static lotto.validation.NumberValidation.validateDuplicateNumber;
+import static lotto.validation.NumberValidation.validateNumberRange;
+import static lotto.validation.NumberValidation.validateNumberSize;
 
 import java.util.Collections;
 import java.util.List;
